requirements.txt
Django>=4.2.11
django-modeltranslation>=0.18.12
Pillow>=10.0.0

